ccid: Fix AIO handling of IN and IRQ endpoint
Even though functionfs forces us to use Linux AIO, we actually don't
want to submit multiple concurrent pendign writes, but we simply want
to emulate the poll/select interface.  So we must introduce a write
(message) queue for the IN and IRQ endpoint, and we must always only
pull one message off the queue, submit the iocb.  On completion, we
dequeue the next message, and so on.

Change-Id: I8f7026f13cf542185cfec4d4780c64965d1706e2