libosmo_emb: Implement _gettimeofday() on simplistic 'jiffies' counter.
We simply use the SysTick timer to count at 1KHz and fill a 64bit jiffies
conunter.  This counter is then used to fill tv_sec and tv_usec in the
gettimeofday() implementation.

NOTE: tv_sec will not indicate the seconds since the epoch (Jan 01,
1970), but rather since system startup.  For the existign users,
particularly osmo_timer, this doesn't matter.

Change-Id: I9dbbb730996bde1e7039f790d76d7243739a8419