relax the timeouts
High baud rates lead to incredibly small wait times that are infeasible
as long as software timers are being used, but this is not really a
problem - they can just be increased to a reasonable value.
At the end of the day the timers need to take the latency of the
firmware into account, and timeouts only occur on powerup with wrong
voltages/dead cards, or in rare error cases.
The actual timeout value does therefore not really matter that much.

Change-Id: Icb733f2d46f45c1a7dcd2abe6bc292bd41d6b9c2