Fix returned data in case of command TPDU
When Lc>0 (command TPDU), our response always copied the command
data in front of the response (which is a SW of 2 bytes in this case).

This of course confuses the hell out of users.  The response to a Lc>0
TPDU should be just the two-bytes status word.

Introduce the use of msgb->l4h for the point in the buffer from where
we receive data from the card.  all bytes before l4h are transmitted
to the card.

Change-Id: Ie3fc7437d39dc330b9f2b7d7960ad2560b6be4b7