usb: reset the device on usb reset if it was previously running
Usb init order is desc read -> reset-> set addr -> use device.

DADD.ADDEN is automatically cleared upon receiving a reset from the
host, at least according to the SAMD51 manual, but DADD.DADD is zero as
well (?) so it can't be used to check if usb was previously running and
the cable was pulled out and plugged in again while the device was
externally powered.
This works around this and ensures a device reset from the host leads to
a proper device restart and reinitialization in this case, there is no
state we want to keep in this case anyway.

Change-Id: I692de3dd98cfc30a010b457be87331e482632fb7
Closes: SYS#4908