ccid fsm: proper wtime calculation
The timeouts had a minor off by a million problem because WT (in ETU) =
WI * 960 as per 7816-3 10.2, but WI was used as WT without the factor of
960 after the pps exchange, so WT (in ETU) was accidentally set to 10
instead of 9600 for the default WI=10.
Additionally that time was only added once, even though the wait time is
the time between the leading character edge of one character and the
next one, irrespective of direction, so the total wait time is this
multiplied with the number of expected bytes.

In practice this had no real effect due to to a reasonable minimum
timeout value and cards that were much faster than this.

Closes: OS#4742
Change-Id: I9ea0b6b51c8cc6f08a36b48e516c64b5e2bbaf9b