7816 fsm: move to static msgb
Error handling was difficult due to the need to pass msgb pointers
attached to the sub fsms back to the cb that is polled from the main
loop to be able free them, while ensuring they never get lost, because
memory leakage is deadly.

This is now fixed by using static pseudo-msgbs for the fsms that are
never deallocated. This only adds one tpdu tx copy, the ccid response
msgbs were already being copied anyway, so memory usage has not changed
except for the "unused slots" case that is not really important since
the octsim was designed around concurrent slot operation anway.

All of this allows convenient error handling in the allstate function
instead of having to spread it all over the sub fsms - in practice
handling errors mostly consists of card deactivation + returning a
proper failure message that matches the ccid command anyway.

Change-Id: I65e77c376aca9ed50e234a0b58a7450a8bbd4fe0