7816 fsm/cuart: support inverse condition cards
This does not currently work with "pcsc-lite version 1.8.26." because
the inverse conditon cards I know do not support the PPS exchange, which
according to spec leads to card deactivation, so the slots needs to be
powered up again, which does not happen.

osmo-sim-test works after manually disabling the pps exchange code in
the firmware.

Change-Id: I892e1d883825111cc1e4ea09589c4fdd256da03c