internal ul32k instead of external 32k osc
This should be fine, the slow clock for the uarts is not used for the
modes we're interested in, and the usb clock is derived from the usb sof
anyway, so it's not used as a reference there either.

This will most likely drift, but so did the external one, and a
free-running rtc without any external input was going to drift anyway,
considering the octsim is supposed to just run without being powered
down for a very long time.

Change-Id: I257b8854e1860ed698c4007efeb5d16e842b519a