Make ch9 usb tests work
This passes all tests except TD 9.9.
That's is weird, because 9.9 works fine when doing
the same commands on linux. It does not really
matter anyway, since stalling the device ep by
sending a message from the host is is only useful
for the test itself... There is no reasonable
way to continue running after random forced ep
stalls anyway, resetting the device would be the
sane option.

This adds some mildly useful debug python snippets
that can be used to inspect the device state.

Change-Id: Icadaee9d8cbe24bd3cac002cc4f710dcf846a32b